import fix

+++ include/services/ConfigService.hpp
@@ -3,12 +3,12 @@
 #include <cstdint>
 struct Config {
     std::string schedulingAlgo;
-    uint32_t quantumCycles;
-    uint32_t cpuCount;
-    uint32_t batchProcessFreq;
-    uint32_t mixIns;
-    uint32_t maxIns;
-    uint32_t delayPerSec;
+    uint64_t quantumCycles;
+    uint64_t cpuCount;
+    uint64_t batchProcessFreq;
+    uint64_t mixIns;
+    uint64_t maxIns;
+    uint64_t delayPerSec;
 };
 
 class ConfigService {
